pyproject.toml
[tool.poetry]
name = "nonebot_plugin_friends"
version = "0.1.4"
description = "Nonebot2插件处理好友申请"
authors = ["Agnes_Digital <Z735803792@163.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/Agnes4m/nonebot_plugin_friends"
repository = "https://github.com/Agnes4m/nonebot_plugin_friends"
keywords = ["maimai", "nonebot2", "plugin"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",

]
include = [
    "LICENSE","README.md"
]
 
[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0"
nonebot-adapter-onebot = "^2.1.3"
[tool.poetry.dev-dependencies]
 
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pyright]
pythonVersion = "3.8"

[tool.black]
line-length = 89
target-version = ["py39", "py310", "py311"]
include = '\.pyi?$'
# skip-string-normalization = true

[tool.ruff.isort]
extra-standard-library = ["typing_extensions"]

[tool.ruff]
ignore = [
    "B008",
    "B905",
    "E501",
    "FBT001",
    "FBT002",
    "FBT003",
    "PGH003",
    "RUF001",
    "RUF002",
    "RUF003",
    "RUF006",
    "RUF100",
    "SIM117",
    "TRY002",
    "TRY003",

]
select = [
    "A",
    "ARG",
    "ASYNC",
    "B",
    "C4",
    "COM",
    "E",
    "F",
    "FBT",
    "FLY",
    "I",
    "ISC",
    "N",
    "PIE",
    "PGH",
    "PTH",
    "PYI",
    "Q",
    "RET",
    "RSE",
    "RUF",
    "SIM",
    "SLF",
    "SLOT",
    "TRY",
]
